+++ src/main/clojure/aerospike_clj/utils.clj
@@ -21,7 +21,8 @@
   (= bin-names [""]))
 
 (defn string-keys?
-  "Predicate function to determine whether all keys provided for bins are strings."
+  {:doc        "Predicate function to determine whether all keys provided for bins are strings."
+   :deprecated "3.1.0"}
   [bin-names]
   (every? string? bin-names))
 
@@ -38,9 +39,19 @@
   (get reverse-boolean-replacements bin-value bin-value))
 
 (defn v->array
-  "An optimized way to convert vectors into Java arrays of type `clazz`."
-  [clazz v]
-  (.toArray ^Collection v ^"[Ljava.lang.Object;" (make-array clazz (count v))))
+  "An optimized way to convert [[java.util.Collection]]s into Java arrays of type `clazz`."
+  ([clazz ^Collection v]
+   (.toArray v ^"[Ljava.lang.Object;" (make-array clazz 0)))
+  ([clazz mapper-fn ^Collection v]
+   (let [size     (.size v)
+         res      ^"[Ljava.lang.Object;" (make-array clazz size)
+         iterator (.iterator v)]
+     (loop [i 0]
+       (when (and (< i size)
+                  (.hasNext iterator))
+         (aset res i (mapper-fn (.next iterator)))
+         (recur (inc i))))
+     res)))
 
 (defn vectorize
   "convert a single value to a vector or any collection to the equivalent vector.