+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [3.1.0] - 2023-08-22
+
+#### Added
+
+* Add the `aerospike-clj.collections/->list` function, which is similar to `clojure.core/mapv`, but it's more efficient
+  when the input is not a Clojure sequence.
+
+#### Changed
+
+* Make the `aerospike-clj.utils/v->array` multi-arity, allowing to pass a `mapper-fn` to map the values before setting
+  them into the array.
+* Optimize the `aerospike-clj.utils/v->array` function by calling `java.util.Collection#toArray` with a 0-length array,
+  this will force the implementation to use the more performant `java.util.Arrays.copyOf`.
+* Add contents: write to the Push CI - master action, this should resolve the git push issues from the GitHub actions
+  bot.
+
+#### Deprecated
+
+* Deprecate `aerospike-clj.utils/string-keys?`.
+
 ## [3.0.0] - 2023-08-03
 
 ### Changed
@@ -200,6 +220,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * License changed to Apache 2.
 
 [A complete list of all java client related changes](https://www.aerospike.com/download/client/java/notes.html)
+
+[3.1.0]: https://github.com/AppsFlyer/aerospike-clj/pull/69
 
 [3.0.0]: https://github.com/AppsFlyer/aerospike-clj/pull/62